+++ game/src/main.ts
@@ -1,0 +1,7 @@
+import StartGame from './game';
+
+document.addEventListener('DOMContentLoaded', () => {
+
+    StartGame('game-container');
+
+});