+++ game/vite/config.dev.mjs
@@ -1,0 +1,18 @@
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+    publicDir: 'assets',
+    base: './',
+    build: {
+        rollupOptions: {
+            output: {
+                manualChunks: {
+                    phaser: ['phaser']
+                }
+            }
+        },
+    },
+    server: {
+        port: 8080
+    }
+});